refactoring to abstract rety logic
