+++ packages/nodejs/src/net/adal/adalfetchclient.ts
