+++ packages/nodejs/src/nodejs.ts
@@ -18,6 +18,6 @@ const NodeFetch = require("node-fetch");
 })(global);
 
 
-export { SPFetchClient, SPOAuthEnv } from "./net/spfetchclient";
-export { SPFetchClientRetry, IRetryData } from "./net/spfetchclientretry";
-export { AdalFetchClient, AADToken } from "./net/adalfetchclient";
+export * from "./net/sharepoint";
+export { AdalFetchClient, AADToken } from "./net/adal/adalfetchclient";
+export { RetryNodeFetchClient } from "./net/retrynodefetchclient";
